% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_cloud_folder.R
\name{backup_cloud_folder}
\alias{backup_cloud_folder}
\title{Backup a Cloud Folder to a Local Zip File}
\usage{
backup_cloud_folder(
  path,
  target_file = get_download_target_file(base_name, backup_dir),
  user = nextcloud_user(),
  auth = nextcloud_auth(),
  backup_dir = kwb.utils::createDirectory("~/../Downloads/nextcloud-backups"),
  base_name = basename(path)
)
}
\arguments{
\item{path}{path to folder on the cloud}

\item{target_file}{path to the target file. A default path is provided by
\code{get_download_target_file(base_name, backup_dir)}}

\item{user}{user name, by default taken from the environment variable
"NEXTCLOUD_USER".}

\item{auth}{authentication header as provided by
\code{kwb.nextcloud:::nextcloud_auth}}

\item{backup_dir}{directory into which to download the zip-file}

\item{base_name}{base name to be integrated into the name of the downloaded
file}
}
\description{
Zip cloud folder and download to local zip file. Default file name:
\code{<backup_dir>/cloud-backup_<base_name>_<yyyy-mm-dd-HHMM>.zip}
}
