% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_file_info_files.R
\name{compare_file_info_files}
\alias{compare_file_info_files}
\title{Compare Two Files Containing File Information}
\usage{
compare_file_info_files(file_1, file_2, ...)
}
\arguments{
\item{file_1}{path to first csv file}

\item{file_2}{path to second csv file}

\item{\dots}{further arguments passed to \code{\link[utils]{read.csv2}}}
}
\description{
Compare two CSV files as retrieved by list_files(..., full_info = TRUE)
}
