% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file_or_folder.R
\name{delete_file_or_folder}
\alias{delete_file_or_folder}
\title{Delete a File or Folder on the Cloud}
\usage{
delete_file_or_folder(
  path,
  user = nextcloud_user(),
  auth = nextcloud_auth(),
  ...
)
}
\arguments{
\item{path}{path to the file or folder to be deleted}

\item{user}{user name, by default taken from the environment variable
"NEXTCLOUD_USER".}

\item{auth}{authentication header as provided by
\code{kwb.nextcloud:::nextcloud_auth}}

\item{\dots}{further arguments passed to
\code{kwb.nextcloud:::nextcloud_request}}
}
\description{
Delete a File or Folder on the Cloud
}
