% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_files}
\alias{download_files}
\title{Download Files from Cloud}
\usage{
download_files(
  hrefs = NULL,
  target_dir = create_download_dir("nextcloud_", dbg = dbg),
  paths = NULL,
  user = nextcloud_user(),
  auth = nextcloud_auth(),
  dbg = TRUE
)
}
\arguments{
\item{hrefs}{href strings specifying the files to be downloaded. The href
strings are returned by \code{\link{list_files}}. Alternatively, the paths
to the files can be given in \code{paths}.}

\item{target_dir}{path to local target directory}

\item{paths}{Alternatively to givin the href strings, the paths to the files
can be given in this argument.}

\item{user}{name of nextcloud user. Default: result of calling
\code{kwb.nextcloud:::nextcloud_user}}

\item{auth}{authentication as returned by
\code{kwb.nextcloud:::nextcloud_user}}

\item{dbg}{logical indicating whether or not to show debug messages}
}
\description{
Download Files from Cloud
}
