% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exists.R
\name{file_exists}
\alias{file_exists}
\title{Does a File Exist on the Nextcloud Server?}
\usage{
file_exists(path)
}
\arguments{
\item{path}{path to the file}
}
\value{
\code{TRUE} if the file exists, else \code{FALSE}
}
\description{
Does a File Exist on the Nextcloud Server?
}
