% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file_versions.R
\name{get_file_versions}
\alias{get_file_versions}
\title{Get Information on Versions of Given Files}
\usage{
get_file_versions(file_info)
}
\arguments{
\item{file_info}{data frame as returned by \code{\link{list_files}} when
being called with \code{full_info = TRUE}}
}
\value{
data frame
}
\description{
Get Information on Versions of Given Files
}
