% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_info.R
\name{get_version_info}
\alias{get_version_info}
\title{Get Version Information on Files of Given IDs}
\usage{
get_version_info(
  file_ids,
  user = nextcloud_user(),
  ignore = "^(status|quota|getcontent|resource|href)"
)
}
\arguments{
\item{file_ids}{vector of integer representing file ids (values returned in
column \code{fileid}) by \code{\link{list_files}}}

\item{user}{nextcloud user}

\item{ignore}{pattern matching columns to be excluded from the result. Set
to \code{NULL} to see what columns are available.}
}
\value{
data frame with one row per version. There seems to be only an entry
if the corresponding file as more than one version.
}
\description{
Get Version Information on Files of Given IDs
}
