% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file_versions.R
\name{list_file_versions}
\alias{list_file_versions}
\title{List the Partner's File Versions (Except Current)}
\usage{
list_file_versions(path, pattern = NULL, ...)
}
\arguments{
\item{path}{relative path to folder of which file versions are to be listed}

\item{pattern}{optional. If specified, only files with names matching the
pattern are considered.}

\item{\dots}{additional arguments passed to \code{\link{list_files}}, such as
\code{recursive}, \code{max_depth}}
}
\description{
List the Partner's File Versions (Except Current)
}
