% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List Files on the Nextcloud Server}
\usage{
list_files(
  path = "",
  pattern = NULL,
  recursive = FALSE,
  full_info = FALSE,
  user = nextcloud_user(),
  password = nextcloud_password(),
  max_depth = NA,
  ...
)
}
\arguments{
\item{path}{path to the nextcloud folder to be listed}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned.}

\item{recursive}{if \code{TRUE} the contents of subfolders are listed as
well, up to the given \code{max_depth}.}

\item{full_info}{if \code{TRUE} the full file information is returned as a
data frame. If \code{FALSE} (the default) only the (relative) file paths
are returned as a vector of character.}

\item{user}{user name, by default taken from the environment variable
"NEXTCLOUD_USER".}

\item{password}{password for nextcloud user, by default taken from the
environment variable "NEXTCLOUD_PASSWORD".}

\item{max_depth}{maximum recursion depth if \code{recursive = TRUE}. By
default \code{max_depth} is \code{NA} meaning that the function behaves
"fully recursive".}

\item{\dots}{further arguments passed to
\code{\link[kwb.utils]{listToDepth}}.}
}
\value{
vector of character or data frame, each with attribute "root" being
set to the value of \code{path}.
}
\description{
List Files on the Nextcloud Server
}
