% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_file_or_folder.R
\name{move_file_or_folder}
\alias{move_file_or_folder}
\title{Move a File or Folder on the Cloud}
\usage{
move_file_or_folder(
  path,
  path_target,
  overwrite = FALSE,
  user = nextcloud_user(),
  auth = nextcloud_auth(),
  dbg = TRUE,
  ...
)
}
\arguments{
\item{path}{path to the file or folder to be moved}

\item{path_target}{path to the target file or folder to be moved to}

\item{overwrite}{overwrite if file/folder is already existing (default: FALSE)}

\item{user}{user name, by default taken from the environment variable
"NEXTCLOUD_USER".}

\item{auth}{authentication header as provided by
\code{kwb.nextcloud:::nextcloud_auth}}

\item{dbg}{print debug messages (default: TRUE)}

\item{\dots}{further arguments passed to
\code{kwb.nextcloud:::nextcloud_request}}
}
\description{
Move a File or Folder on the Cloud
}
