% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_to_file_href}
\alias{path_to_file_href}
\title{Convert Path to href as Required for Download}
\usage{
path_to_file_href(path = "", user = nextcloud_user(), leading_slash = FALSE)
}
\arguments{
\item{path}{path to file on Nextcloud server}

\item{user}{name of Nextcloud user, default:
\code{kwb.nextcloud:::nextcloud_user()}}

\item{leading_slash}{should the returned href start with a slash?}
}
\description{
Convert Path to href as Required for Download
}
