% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\title{Upload a Local File to the Cloud}
\usage{
upload_file(
  file,
  target_path = "documents",
  user = nextcloud_user(),
  auth = nextcloud_auth()
)
}
\arguments{
\item{file}{path to local file}

\item{target_path}{path to folder on the cloud into which to upload the file}

\item{user}{user name, by default taken from the environment variable
"NEXTCLOUD_USER".}

\item{auth}{authentication header as provided by
\code{kwb.nextcloud:::nextcloud_auth}}
}
\description{
Upload a Local File to the Cloud
}
